api/requirements.txt
django==4.2.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
django-debug-toolbar
Pillow==9.5.0
djangorestframework==3.14.0
markdown==3.4.3
django-filter==23.2
django-cors-headers==4.1.0